% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joining-metadata.R
\name{add_missing_specimens}
\alias{add_missing_specimens}
\title{Add missing specimenID rows}
\usage{
add_missing_specimens(meta_df, annot_df)
}
\arguments{
\item{meta_df}{Dataframe with columns \code{specimenID}, \code{individualID}, \code{assay},
\code{study}.}

\item{annot_df}{Dataframe with columns \code{specimenID}, \code{individualID}, \code{assay},
\code{study}.}
}
\description{
Add missing \code{specimenID}s from \code{annot_df} to \code{meta_df}. Groups
the check for missing \code{specimenIDs} by \code{assay} and \code{study}.
}
