% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locating-metadata-files.R
\name{child_names}
\alias{child_names}
\title{Pull out names of children}
\usage{
child_names(dir_id)
}
\arguments{
\item{dir_id}{Synapse folder synID}
}
\description{
Get the names of the child folders.
}
