% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cleaning_utils.R
\name{clean_json_string}
\alias{clean_json_string}
\title{Remove JSON From Multi-annotations}
\usage{
clean_json_string(json_string, remove_spaces = FALSE)
}
\arguments{
\item{json_string}{Character string with JSON elements.}

\item{remove_spaces}{\code{TRUE} to also remove spaces, else \code{FALSE}. Note that
this will remove \emph{all} spaces. Default is \code{FALSE}.}
}
\value{
Character string without brackets, escaped quotes, and (optionally)
spaces.
}
\description{
Remove the JSON-specific elements from a string. For example,
a JSON string could look like \verb{"[\\"foo\\", \\"bar\\"]"}. This function would
return the string \code{"foo, bar"}.
}
