% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering-ids-from-metadata-files.R
\name{gather_ids_all_studies}
\alias{gather_ids_all_studies}
\title{Gather IDs from metadata files}
\usage{
gather_ids_all_studies(all_files)
}
\arguments{
\item{all_files}{dataframe with columns \code{id} (synID of file), \code{study} (name
of related study), \code{metadataType} (\code{assay}, \code{individual}, or \code{biospecimen}),
\code{assay} (type of assay), \code{dataType} (type of data).}
}
\value{
Dataframe with columns \code{individualID}, \code{specimenID}, \code{assay},
\code{study} containing all relevant metadata from \code{all_files}.
}
\description{
Assumes all metadata files are csv with some combination of
columns: \code{individualID}, \code{specimenID}, \code{assay}. Gathers by study grouping.
}
