% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering-ids-from-metadata-files.R
\name{gather_ids_by_study}
\alias{gather_ids_by_study}
\title{Gather study metadata IDs}
\usage{
gather_ids_by_study(meta_files)
}
\arguments{
\item{meta_files}{Dataframe with columns \code{metadataType} (\code{assay},
\code{individual}, or \code{biospecimen}), \code{assay} (type of assay), and \code{data} (nested
dataframe with the metadata from the file).}
}
\description{
Gather metadata IDs for a single study.
}
