% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locating-metadata-files.R
\name{gather_metadata_synIDs_all}
\alias{gather_metadata_synIDs_all}
\title{Gather Metadata synIDs}
\usage{
gather_metadata_synIDs_all(dir_id)
}
\arguments{
\item{dir_id}{Synapse synID for top level directory.}
}
\value{
Dataframe with file synID, filename, and study name.
}
\description{
Gather synIDs for metadata files. Assumes all children of
\code{dir_id} are named study folders.
}
