% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering-ids-from-metadata-files.R
\name{get_all_study_metadata}
\alias{get_all_study_metadata}
\title{Get All Study Data}
\usage{
get_all_study_metadata(
  meta_files,
  columns = c("individualID", "specimenID", "assay")
)
}
\arguments{
\item{meta_files}{dataframe with columns \code{id} (synID of file), \code{study} (name
of related study), \code{metadataType} (\code{assay}, \code{individual}, or \code{biospecimen}),
\code{assay} (type of assay), \code{dataType} (type of data).}

\item{columns}{Vector of column names to keep in the data. Discards all other
metadata columns. Default is \code{c("individualID", "specimenID", "assay")}.}
}
\value{
\code{meta_files} with additional list column called \code{data} with the data
from each file.
}
\description{
Add a nested dataframe column that contains only the desired
data from each metadata file.
}
