% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script-helpers.R
\name{get_config}
\alias{get_config}
\title{Get config}
\usage{
get_config(value, config)
}
\arguments{
\item{value}{the config value to return}

\item{config}{the config settings needed}
}
\description{
Get config
}
