% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering-ids-from-metadata-files.R
\name{get_file_data}
\alias{get_file_data}
\title{Get File Data}
\usage{
get_file_data(syn_id)
}
\arguments{
\item{syn_id}{Synapse synID for a file.}
}
\value{
Dataframe with the file data or \code{NA} if the file could not be
opened.
}
\description{
Download file from Synapse and get data from it.
}
