% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering-ids-from-metadata-files.R
\name{get_file_indices}
\alias{get_file_indices}
\title{Get Indices for All Metadata File Types}
\usage{
get_file_indices(
  meta_files,
  meta_types = c("biospecimen", "assay", "individual")
)
}
\arguments{
\item{meta_files}{Dataframe with column \code{metadataType} (type of metadata).}

\item{meta_types}{Vector of metadata types to get indices for. Default is
\code{c("biospecimen", "assay", "individual")}.}
}
\description{
Get a list of file indices for each metadata file type.
}
