% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-tables.R
\name{simple_table_query}
\alias{simple_table_query}
\title{Get Simple Table Query}
\usage{
simple_table_query(
  table_id,
  columns = c("study", "individualID", "specimenID", "assay")
)
}
\arguments{
\item{table_id}{Synapse synID of table to query.}

\item{columns}{Vector of desired columns from the table. Default is
\code{c("study", "individualID", "specimenID", "assay")}.}
}
\value{
Dataframe from query with only the requested columns.
}
\description{
Query a Synapse table for specific columns and only receive
those columns back in dataframe.
}
