% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-tables.R
\name{update_samples_table}
\alias{update_samples_table}
\title{Update sample ID table}
\usage{
update_samples_table(table_id, new_data)
}
\arguments{
\item{table_id}{synID of the table to overwrite.}

\item{new_data}{the new data to overwrite the table with.}
}
\description{
\strong{Warning} This is a destructive function!
Update the sample ID table with new data. This function will delete the rows
in the old data and upload the new data in its place. The existing table
schema must match the columns in \code{new_data}.
}
