% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script-helpers.R
\name{update_task_annotation}
\alias{update_task_annotation}
\title{Update task annotation}
\usage{
update_task_annotation(task_id, annots, success, task_view = NA)
}
\arguments{
\item{annots}{Synapse annotation object for the task folder}

\item{success}{TRUE if the task was completed successfully, else FALSE}

\item{task_view}{synID for the task file view. If provided, will update the
view after annotating the task folder.}
}
\description{
Update the 'completed_successfully' task annotation on a Synapse folder.
Expected that the annotation is a text type.
}
