% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script-helpers.R
\name{upload_log}
\alias{upload_log}
\title{Upload the log file to Synapse}
\usage{
upload_log(folder, path)
}
\arguments{
\item{folder}{synID of parent folder for logs}

\item{path}{log file path}
}
\description{
Upload the log file to Synapse
}
